+++ CHANGES.rst
@@ -5,3 +5,5 @@ CHANGES
 ------------------
 
 - Dropped: `aiohttp.protocol.HttpPrefixParser`  #1590
+
+- Dropped: Servers response's `.started`, `.start()` and `.can_start()` method  #1591